refactor(joint-react): update imports to use cell.types for improved type consistency
