+++ packages/joint-react/src/hooks/__tests__/use-cell.type.test.ts
@@ -32,7 +32,7 @@ import type {
   LinkRecord,
   CellRecord,
 } from '../../types/cell.types';
-import type { ElementPosition } from '../../types/cell-data';
+import type { ElementPosition } from '../../types/cell.types';
 
 interface ElementUserData {
   readonly label: string;
